requirements.txt

openai
streamlit
streamlit-chat
pyyaml
jsonschema
load_dotenv
